Suppression limitation de la taille des tableaux pour le détail insertions (conflits entre javascripts)
git-tfs-id: [http://172.17.236.41:8080/]$/AdExpress International Demo/Trunk;C7195

+++ DLL/TNS.AdExpressI.Portofolio.Poland/Engines/SynthesisEngine.cs
@@ -245,8 +245,8 @@ namespace TNS.AdExpressI.Portofolio.Poland.Engines {
             #endregion
 
             #region nbLines init
-            long nbLines = 0;
-            long lineIndex = 0;
+            Int32 nbLines = 0;
+            Int32 lineIndex = 0;
             switch (_vehicleInformation.Id) {
                 case TNS.AdExpress.Constantes.Classification.DB.Vehicles.names.press:
                     nbLines = 9;
